+++ src/main/java/http/HttpSession.java
@@ -1,12 +1,58 @@
 package http;
 
+import java.util.LinkedHashMap;
+import java.util.Map;
+import java.util.Objects;
+import java.util.UUID;
+
 /**
  * @author KingCjy
  */
-public interface HttpSession {
-    String getId();
-    void setAttribute(String name, Object value);
-    Object getAttribute(String name);
-    void removeAttribute(String name);
-    void invalidate();
+public class HttpSession {
+
+    private Map<String, Object> attributes = new LinkedHashMap<>();
+
+    private String id;
+
+    HttpSession() {
+        this(UUID.randomUUID().toString());
+    }
+
+    HttpSession(String id) {
+        this.id = id;
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public void setAttribute(String name, Object value) {
+        attributes.put(name, value);
+    }
+
+    public Object getAttribute(String name) {
+        return attributes.get(name);
+    }
+
+    public void removeAttribute(String name) {
+        attributes.remove(name);
+    }
+
+    public void invalidate() {
+        attributes.clear();
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) return true;
+        if (o == null || getClass() != o.getClass()) return false;
+        if (!super.equals(o)) return false;
+        HttpSession that = (HttpSession) o;
+        return Objects.equals(id, that.id);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(super.hashCode(), id);
+    }
 }
