--- src/main/java/http/HttpSessionImpl.java
@@ -1,63 +1,0 @@
-package http;
-
-import java.util.LinkedHashMap;
-import java.util.Map;
-import java.util.Objects;
-import java.util.UUID;
-
-/**
- * @author KingCjy
- */
-public class HttpSessionImpl implements HttpSession {
-
-    private Map<String, Object> attributes = new LinkedHashMap<>();
-
-    private String id;
-
-    HttpSessionImpl() {
-        this(UUID.randomUUID().toString());
-    }
-
-    HttpSessionImpl(String id) {
-        this.id = id;
-    }
-
-    @Override
-    public String getId() {
-        return id;
-    }
-
-    @Override
-    public void setAttribute(String name, Object value) {
-        attributes.put(name, value);
-    }
-
-    @Override
-    public Object getAttribute(String name) {
-        return attributes.get(name);
-    }
-
-    @Override
-    public void removeAttribute(String name) {
-        attributes.remove(name);
-    }
-
-    @Override
-    public void invalidate() {
-        attributes.clear();
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        if (!super.equals(o)) return false;
-        HttpSessionImpl that = (HttpSessionImpl) o;
-        return Objects.equals(id, that.id);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(super.hashCode(), id);
-    }
-}
